% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_Dummy.R
\name{Add_Dummy}
\alias{Add_Dummy}
\title{Dummy Variables}
\usage{
Add_Dummy(mod_obj)
}
\arguments{
\item{mod_obj}{- model object}
}
\value{
mod_obj
}
\description{
This adds dummy variables to data and modifies the spec
}
