% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_Dummy3.R
\name{Add_Dummy3}
\alias{Add_Dummy3}
\title{Dummy Variables}
\usage{
Add_Dummy3(mod_obj)
}
\arguments{
\item{mod_obj}{- model object}
}
\value{
mod_obj
}
\description{
This adds dummy variables to data and modifies the spec
}
