% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_group_selector.R
\name{add_group_selector}
\alias{add_group_selector}
\title{Add Group Selector}
\usage{
add_group_selector(mod_obj, vehicles, regions)
}
\arguments{
\item{mod_obj}{- model object}

\item{vehicles}{- cahacter vector}

\item{regions}{- character vectors}
}
\description{
Add Group Selector
}
