% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottom_up_decomp.R
\name{bottom_up_decomp}
\alias{bottom_up_decomp}
\title{Bottom-Up Decomp}
\usage{
bottom_up_decomp(mod_obj, min_ref_var_names = NULL, mean_ref_var_names = NULL)
}
\arguments{
\item{mod_obj}{- model object}

\item{min_ref_var_names}{- vector of variable indexes}

\item{mean_ref_var_names}{- vector of variable indexes}
}
\value{
mod_obj
}
\description{
This Decomp function is compatible with the Bayesian models fitted with the MSMP framework
The format of the output is compatible with the legacy reach generation function.
}
