% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_setup.R
\name{create_model_setup}
\alias{create_model_setup}
\title{Create Model Setup}
\usage{
create_model_setup(
  ModelForm = "lin_lin",
  Panel = "Y",
  Crossection = NULL,
  Time = "month",
  BeginDate = NULL,
  EndDate = NULL,
  kpi = NULL,
  nmp = NULL,
  rgn = NULL
)
}
\value{
model_setup_table
}
\description{
This function creates a model setup table data.frame
}
