% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_spec.R
\name{create_model_spec}
\alias{create_model_spec}
\title{Create Model Spect}
\usage{
create_model_spec(mod_obj)
}
\value{
model_spec_table
}
\description{
This function creates a model spec template data.frame
}
