% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decomp.R
\name{Decomp}
\alias{Decomp}
\title{Model Decomposition and Contribution Calculations}
\usage{
Decomp(mod_obj, min_ref = NULL, max_ref = NULL, mean_ref = NULL)
}
\arguments{
\item{mod_obj}{- model object}

\item{min_ref}{- vector of variable indexes}

\item{max_ref}{- vector of variable indexes}

\item{mean_ref}{- vector of variable indexes}

\item{kpi}{- kpi variable string}
}
\value{
mod_obj
}
\description{
This Decomp function is compatible with the Bayesian models fitted with the MSMP framework
The format of the output is compatible with the legacy reach generation function.
}
