% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decomp2.R
\name{Decomp2}
\alias{Decomp2}
\title{Model Decomposition and Contribution Calculations}
\usage{
Decomp2(mod_obj)
}
\arguments{
\item{mod_obj}{- model object}

\item{min_ref}{- vector of variable indexes}

\item{max_ref}{- vector of variable indexes}

\item{mean_ref}{- vector of variable indexes}
}
\value{
mod_obj
}
\description{
This Decomp function is compatible with the Bayesian models fitted with the MSMP framework
The format of the output is compatible with the legacy reach generation function.
}
