% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decomp3.R
\name{Decomp3}
\alias{Decomp3}
\title{Model Decomposition and Contribution Calculations}
\usage{
Decomp3(mod_obj, min_ref_var_names = NULL, mean_ref_var_names = NULL)
}
\arguments{
\item{mod_obj}{- model object}

\item{min_ref_var_names}{- vector of variable indexes}

\item{mean_ref_var_names}{- vector of variable indexes}
}
\value{
mod_obj
}
\description{
This Decomp function is compatible with the Bayesian models fitted with the MSMP framework
The format of the output is compatible with the legacy reach generation function.
}
