% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_grps_aggregate.R
\name{effective_grps_aggregate}
\alias{effective_grps_aggregate}
\title{effective_grps_aggregate}
\usage{
effective_grps_aggregate(grps_vec, begin, window, decay, simple = TRUE)
}
\arguments{
\item{grps_vec}{- grps_vec}

\item{begin}{- begin}

\item{window}{- window}

\item{decay}{- decay}

\item{simple}{- simple}
}
\description{
effective_grps_aggregate
}
