% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_contribution_curve.R
\name{generate_contribution_curve}
\alias{generate_contribution_curve}
\title{generate_contribution_curve}
\usage{
generate_contribution_curve(reach_curve_spec_row, reach_curve)
}
\arguments{
\item{reach_curve}{}
}
\description{
generate_contribution_curve
}
