% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alpha_beta_gamma.R
\name{get_alpha_beta_gamma}
\alias{get_alpha_beta_gamma}
\title{get_alpha_beta_gamma}
\usage{
get_alpha_beta_gamma(channel_name, fit_curves)
}
\arguments{
\item{channel_name}{- channel_name}

\item{fit_curves}{- fit_curves}
}
\description{
get_alpha_beta_gamma
}
