% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_channel_reach.R
\name{get_channel_reach}
\alias{get_channel_reach}
\title{get_channel_reach}
\usage{
get_channel_reach(E, R, eGRPs, teGRPs, alpha_beta_gamma)
}
\arguments{
\item{obj}{- mod_obj}
}
\description{
get_channel_reach
}
