% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reach_tables.R
\name{get_reach_tables}
\alias{get_reach_tables}
\title{get_reach_tables}
\usage{
get_reach_tables(grps_split, E, R, window, decay, alpha_beta_gamma)
}
\arguments{
\item{grps_split}{- grps_split}

\item{E}{- E}

\item{R}{- R}

\item{window}{- window}

\item{decay}{- decay}

\item{alpha_beta_gamma}{- alpha_beta_gamma}
}
\description{
get_reach_tables
}
