% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.data_transformed.R
\name{is.data_transformed}
\alias{is.data_transformed}
\title{Test data_transformed mod_obj}
\usage{
is.data_transformed(x)
}
\arguments{
\item{x}{- mod_obj}
}
\description{
Test data_transformed mod_obj
}
