% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_Data2.R
\name{Load_Data}
\alias{Load_Data}
\title{Load_Data}
\usage{
Load_Data(
  obj,
  input_files = list(input_file_ModelData = NULL, input_file_FMIData = NULL,
    input_file_SpendData = NULL, input_file_MSRPData = NULL),
  NAMEPLATE = NULL,
  nmp = NULL
)
}
\arguments{
\item{obj}{- model object}

\item{nmp}{}
}
\description{
Load_Data
}
