% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_FitCurves.R
\name{Load_FitCurves}
\alias{Load_FitCurves}
\title{Load_FitCurves}
\usage{
Load_FitCurves(obj, input_file_ModelFitCurves)
}
\arguments{
\item{obj}{- model object}

\item{input_file_ModelFitCurves}{- file path}
}
\description{
Load_FitCurves
}
