% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_FMIData.R
\name{Load_FMIData}
\alias{Load_FMIData}
\title{Load_FMIData}
\usage{
Load_FMIData(obj, input_file_FMIData)
}
\arguments{
\item{obj}{- model object}

\item{input_file_FMIData}{}
}
\description{
Load_FMIData
}
