% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_ModelData.R
\name{Load_ModelData}
\alias{Load_ModelData}
\title{Load_ModelData}
\usage{
Load_ModelData(mod_obj, input_file_ModelData)
}
\arguments{
\item{mod_obj}{- model mod_object}

\item{input_file_ModelData}{- path to an .xlsx file with monthly and weekly data tabs.}
}
\value{
mod_obj
}
\description{
This function load the monthly and weekly level KPI and FMI data.
}
