% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_MSRPData.R
\name{Load_MSRPData}
\alias{Load_MSRPData}
\title{Load_FMIData}
\usage{
Load_MSRPData(obj, input_file_MSRPData, NAMEPLATE = NULL)
}
\arguments{
\item{obj}{- model object}

\item{NAMEPLATE}{}
}
\description{
Load_FMIData
}
