% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_mix_opt.R
\name{media_mix_opt}
\alias{media_mix_opt}
\title{media_mix_op}
\usage{
media_mix_opt(
  mod_obj,
  total_budget = NULL,
  incremental = 1e+05,
  interactive = FALSE
)
}
\arguments{
\item{mod_obj}{- model object}

\item{total_budget}{- total budget}

\item{interactive}{}
}
\description{
media_mix_op
}
