% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmmactive-package.R
\docType{package}
\name{mmmactive}
\alias{mmmactive}
\title{mmmactive: A package for computating the notorious bar statistic.}
\description{
mmmactive package: browseVignettes(package = "mmmactive")
}
\section{Core Idea}{

some desc
}

