% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Resids.R
\name{Plot_Resids}
\alias{Plot_Resids}
\title{Fitted vs Actual and Residual Plot}
\usage{
Plot_Resids(mod_obj, panel_name)
}
\arguments{
\item{mod_obj}{- model mod_object}

\item{panel_name}{- panel name}
}
\value{
mod_obj
}
\description{
A Plot of fitted vs observed for a mod_obj
}
