% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_Contrib1.R
\name{prep_Contrib1}
\alias{prep_Contrib1}
\title{prep_Contrib1}
\usage{
prep_Contrib1(Contrib, spend, year, panel)
}
\arguments{
\item{Contrib}{- Contrib}

\item{spend}{- spend}

\item{year}{- year}

\item{panel}{- panel}
}
\description{
prep_Contrib1
}
