% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{prep_data}
\usage{
prep_data(fmi_data, From, To, Allparams, Contrib1, panel)
}
\arguments{
\item{fmi_data}{- fmi_data}

\item{From}{- From}

\item{To}{- To}

\item{Allparams}{- Allparams}

\item{Contrib1}{- Contrib1}

\item{panel}{- panel}
}
\description{
prep_data
}
