% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReachCurveCalc2.R
\name{ReachCurveCalc2}
\alias{ReachCurveCalc2}
\title{ReachCurveCalc2}
\usage{
ReachCurveCalc2(Values, input, fit_curves)
}
\arguments{
\item{Values}{- Values}

\item{input}{- input}

\item{fit_curves}{- fit_curves}
}
\description{
ReachCurveCalc2
}
