% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReachCurveCalc3.R
\name{ReachCurveCalc3}
\alias{ReachCurveCalc3}
\title{ReachCurveCalc3}
\usage{
ReachCurveCalc3(reach_curve_spec, grps_data, fit_curves)
}
\arguments{
\item{fit_curves}{}
}
\description{
ReachCurveCalc3
}
