% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Response_Curves.R
\name{Response_Curves}
\alias{Response_Curves}
\title{Response_Curves}
\usage{
Response_Curves(mod_obj, Contrib, year)
}
\arguments{
\item{Contrib}{- Contrib}

\item{year}{- year}

\item{obj}{- mod_obj}
}
\description{
Response_Curves
}
