% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Model.R
\name{Run_Model}
\alias{Run_Model}
\title{Execute Model Fitting}
\usage{
Run_Model(mod_obj, method = "linear_regression")
}
\arguments{
\item{mod_obj}{- model object}

\item{Method}{- Fitting method}
}
\value{
mod_obj
}
\description{
Fitting an MMM model on top of a mod_obj
}
