% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Model3.R
\name{Run_Model3}
\alias{Run_Model3}
\title{Execute Model Fitting}
\usage{
Run_Model3(mod_obj, method = NULL, default_priors = FALSE)
}
\arguments{
\item{mod_obj}{- model object}

\item{Method}{- Fitting method}
}
\value{
mod_obj
}
\description{
Fitting an MMM model on top of a mod_obj
}
