% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Model_Panel.R
\name{Run_Model_Panel}
\alias{Run_Model_Panel}
\title{Execute Model Fitting}
\usage{
Run_Model_Panel(mod_obj, Method = "Bayes")
}
\arguments{
\item{mod_obj}{- model object}

\item{Method}{- Fitting method}
}
\value{
mod_obj
}
\description{
Fitting an MMM model on top of a mod_obj
}
\examples{

}
