% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_Date_Range.R
\name{Set_Date_Range}
\alias{Set_Date_Range}
\title{Set_Date_Range}
\usage{
Set_Date_Range(mod_obj, na.rm = TRUE)
}
\arguments{
\item{mod_obj}{- model object}

\item{na.rm}{- logical}
}
\value{
mod_obj
}
\description{
Set the data range for the mod_obj before model fitting
Uses the BeginData and EndDate attributes of the mod_obj to set the range on the transformed data.
}
