% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform.R
\name{Transform}
\alias{Transform}
\title{Transform}
\usage{
Transform(mod_obj, print = TRUE)
}
\arguments{
\item{mod_obj}{- model mod_object}

\item{print}{- TRUE or FALSE}
}
\value{
mod_obj
}
\description{
Run data transformation and combine weekly and monthly level data accosting to the model specification.
}
