% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform2.R
\name{Transform2}
\alias{Transform2}
\title{Transform2}
\usage{
Transform2(mod_obj, print = TRUE)
}
\arguments{
\item{mod_obj}{- model mod_object}

\item{print}{- TRUE or FALSE}
}
\value{
mod_obj
}
\description{
Run data transformation and combine weekly and monthly level data accosting to the model specification.
}
