% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformSplit.R
\name{TransformSplit}
\alias{TransformSplit}
\title{TransformSplit}
\usage{
TransformSplit(
  data_split = NULL,
  spec_split = NULL,
  fit_curves = NULL,
  print = TRUE
)
}
\arguments{
\item{print}{}
}
\value{
data.frame
}
\description{
TransformSplit
}
