% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformTemp.R
\name{TransformTemp}
\alias{TransformTemp}
\title{TransformTemp}
\usage{
TransformTemp(
  data_tmp = NULL,
  spec_tmp = NULL,
  tmp = NULL,
  fit_curves = NULL,
  cross_section = NULL,
  print = TRUE
)
}
\arguments{
\item{print}{}
}
\description{
TransformTemp
}
