% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformVar.R
\name{TransformVar}
\alias{TransformVar}
\title{TransformVar}
\usage{
TransformVar(
  data_vector = NULL,
  spec_row = NULL,
  fit_curves = NULL,
  print = TRUE
)
}
\arguments{
\item{data_vector}{- vector}

\item{spec_row}{- row from spec table}

\item{fit_curves}{- fit curves}

\item{print}{- logical}
}
\value{
vector - transformed data_vector
}
\description{
TransformVar
}
