% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unnest.R
\name{Unnest}
\alias{Unnest}
\title{Model Decomp Unnesting}
\usage{
Unnest(sales_mod_obj, sub_mod_obj)
}
\arguments{
\item{sales_mod_obj}{- model object}

\item{sub_mod_obj}{- model object}
}
\value{
mod_obj
}
\description{
Unnesting functionality
}
