% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_unwind.R
\name{unwind}
\alias{unwind}
\title{unwind}
\usage{
unwind(
  model1,
  model2,
  use_model2_panel = "ALL",
  print_summary = TRUE,
  expandPeriodToLongest = FALSE,
  demo = FALSE
)
}
\arguments{
\item{model1}{- model1}

\item{model2}{- model2}

\item{use_model2_panel}{- use_model2_panel}

\item{print_summary}{- print_summary}

\item{expandPeriodToLongest}{- logical}

\item{demo}{- logical}
}
\description{
unwind
}
