% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_model_spec.R
\name{update_model_spec}
\alias{update_model_spec}
\title{Update Model Setup}
\usage{
update_model_spec(
  model_spec_table,
  var,
  Trans_Variable = NULL,
  AggregateVariable = NULL,
  Variable_Type = NULL,
  Include = NULL,
  VaryBy = NULL,
  Transform = NULL,
  TransformType = NULL,
  Lag = NULL,
  Scale = NULL,
  Effective = NULL,
  Recency = NULL,
  Decay = NULL,
  Period = NULL,
  Alpha = NULL,
  Window = NULL,
  Prior_Mean = NULL,
  Prior_SD = NULL,
  PriorSD_Adj = NULL,
  Sign = NULL,
  Override = NULL
)
}
\value{
model_setup_table
}
\description{
This function updates a model spec for a given variable row
}
