% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Var_Test_Cor.R
\name{Var_Test_Cor}
\alias{Var_Test_Cor}
\title{Var_Test_Cor}
\usage{
Var_Test_Cor(mod_obj = NULL, print = NULL)
}
\arguments{
\item{mod_obj}{- model mod_object}

\item{print}{- TRUE or FALSE}
}
\value{
data.frame
}
\description{
Test Corr vs Resids
}
