% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate_dummy_spec.R
\name{activate_dummy_spec}
\alias{activate_dummy_spec}
\title{Activate Dummy Specification}
\usage{
activate_dummy_spec(
  mod_obj,
  input_file_DummySpec = NULL,
  dummy_spec_table = NULL
)
}
\arguments{
\item{mod_obj}{- model object}

\item{input_file_ModelSpec}{- file path to ModelSpec.csv}

\item{model_spec_table}{- model_spec_table}
}
\value{
mod_obj
}
\description{
This function onloads the dummy variable specification
}
