% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate_model_setup.R
\name{activate_model_setup}
\alias{activate_model_setup}
\title{Activate Model Setup}
\usage{
activate_model_setup(mod_obj, input_file_ModelSetup)
}
\arguments{
\item{mod_obj}{- model object}

\item{input_file_ModelSetup}{- file path to ModelSetup.csv}
}
\value{
mod_obj
}
\description{
This function dispatches a number of model metadata onto the mod_obj.
}
