% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate_model_spec.R
\name{activate_model_spec}
\alias{activate_model_spec}
\title{Activate Model Specification}
\usage{
activate_model_spec(mod_obj, input_file_ModelSpec)
}
\arguments{
\item{mod_obj}{- model object}

\item{input_file_ModelSpec}{- file path to ModelSpec.csv}
}
\value{
mod_obj
}
\description{
This function onloads the model specification file and includes only modeled variables
}
