% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_Data.R
\name{Load_Data}
\alias{Load_Data}
\title{Load_Data}
\usage{
Load_Data(obj, input_file_ModelData)
}
\arguments{
\item{obj}{- model object}

\item{input_file_ModelData}{- file parth}
}
\description{
Load_Data
}
