% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_Data.R
\name{Load_Data}
\alias{Load_Data}
\title{Load_Data}
\usage{
Load_Data(
  mod_obj,
  input_files = list(input_file_ModelData = NULL, input_file_FMIData = NULL,
    input_file_SpendData = NULL, input_file_MSRPData = NULL),
  NAMEPLATE = NULL,
  nmp = NULL
)
}
\arguments{
\item{mod_obj}{- model mod_object}

\item{input_files}{- paths to input files}

\item{NAMEPLATE}{- Nameplate (ex. Titan)}

\item{nmp}{- Nameplate (ex. ttn)}
}
\value{
mod_obj
}
\description{
This function loads data onto a mod_obj for the paths supplied with the `input_files` vector of paths.
Internally the Load_ModelData(), Load_FMIData(), Load_MSRPData(), and Load_SpendData() functions are called.
}
