% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_FitCurves.R
\name{Load_FitCurves}
\alias{Load_FitCurves}
\title{Load_FitCurves}
\usage{
Load_FitCurves(mod_obj, input_file_ModelFitCurves)
}
\arguments{
\item{mod_obj}{- model mod_object}

\item{input_file_ModelFitCurves}{- file path}
}
\value{
mod_obj
}
\description{
Loads a list of fitcurves onto the mod_obj
}
