% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_FMIData.R
\name{Load_FMIData}
\alias{Load_FMIData}
\title{Load_FMIData}
\usage{
Load_FMIData(mod_obj, input_file_FMIData)
}
\arguments{
\item{mod_obj}{- model mod_object}

\item{input_file_FMIData}{- path to FMI data file}
}
\value{
mod_obj
}
\description{
This function loads FMI data onto a mod_obj.
}
