% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_MSRPData.R
\name{Load_MSRPData}
\alias{Load_MSRPData}
\title{Load_MSRPData}
\usage{
Load_MSRPData(mod_obj, input_file_MSRPData)
}
\arguments{
\item{mod_obj}{- model mod_object}

\item{input_file_MSRPData}{- path to MSRP data file.}
}
\value{
mod_obj
}
\description{
This function loads the MSRP data onto the mod_obj
}
