% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_SpendData.R
\name{Load_SpendData}
\alias{Load_SpendData}
\title{Load_SpendData}
\usage{
Load_SpendData(obj, input_file_SpendData, NAMEPLATE = NULL, nmp = NULL)
}
\arguments{
\item{obj}{- model object}

\item{nmp}{}
}
\description{
Load_SpendData
}
