% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_SpendData.R
\name{Load_SpendData}
\alias{Load_SpendData}
\title{Load_SpendData}
\usage{
Load_SpendData(mod_obj, input_file_SpendData)
}
\arguments{
\item{mod_obj}{- model mod_object}

\item{input_file_SpendData}{- path to the spend data file}
}
\value{
mod_obj
}
\description{
Loads the spend data onto the mod_obj.
}
