% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmmactive-package.R
\docType{package}
\name{mmmactive}
\alias{mmmactive}
\title{mmmactive: Implementation of MSMP framework for MMM workflows.}
\description{
This package implement a set of functions for MMM projects following the MSMP (Marketing Science Modeling Platform) framework.
These functions are designed around a data structure of class `mod_obj` which is generally the first argument. The core functions of the
package are named after a key step of an MMM workflow.
}
\details{
Getting familiar with the `mod_obj` : vignette("mod_obj_class_vignette", "mmmactive")

\enumerate{
      \item \code{Load_Data} : vignette("LoadingData_vignette", "mmmactive")
      \item \code{Transform} : vignette("Transform_vignette", "mmmactive")
      \item \code{Run_Model_Panel} : vignette("Modeling_vignette", "mmmactive")
      \item \code{Decomp + Unnest}
      \item \code{Response}
      \item \code{Media Mix Optimization}
}
}
\section{Vignettes}{

mmmactive package: browseVignettes(package = "mmmactive")
}

\author{
\itemize{
     \item Ben Denis Shaffer (Maintainer),
     \item Mansi Sheth,
     \item Aiden Hu
     \item Julia Liu
}
}
