% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformSplit.R
\name{TransformSplit}
\alias{TransformSplit}
\title{TransformSplit}
\usage{
TransformSplit(
  data_split = NULL,
  spec_split = NULL,
  fit_curves = NULL,
  print = TRUE
)
}
\arguments{
\item{data_split}{- data.frame, slice of a data.}

\item{spec_split}{- spec}

\item{fit_curves}{- fit curves}

\item{print}{- logical}
}
\value{
data.frame
}
\description{
TransformSplit
}
