% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformTemp.R
\name{TransformTemp}
\alias{TransformTemp}
\title{TransformTemp}
\usage{
TransformTemp(
  data_tmp = NULL,
  spec_tmp = NULL,
  tmp = NULL,
  fit_curves = NULL,
  cross_section = NULL,
  print = TRUE
)
}
\arguments{
\item{data_tmp}{- weekly or monthly input data}

\item{spec_tmp}{- spec}

\item{tmp}{- time agg level}

\item{fit_curves}{- fit curves}

\item{cross_section}{- cross_section level}

\item{print}{- logical}
}
\value{
data.frame
}
\description{
TransformTemp
}
