% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformTempJoin.R
\name{TransformTempJoin}
\alias{TransformTempJoin}
\title{TransformTempJoin}
\usage{
TransformTempJoin(data_input, DFFS, trans_variable)
}
\arguments{
\item{data_input}{- list of weekly and monthly data.frames}

\item{DFFS}{- list of split data.frames}

\item{trans_variable}{- variable}
}
\description{
TransformTempJoin
}
